python/requirements.txt
pymongo==3.12.0
pandas==1.2.5
tweepy==3.10.0
beautifulsoup4==4.9.3
lxml==4.6.3
ipython==7.25.0
python-dotenv==0.19.1